refactor: remove command-layer cruft (test_scope.rs, refactor_tests.rs)
test_scope.rs was a dead stub — declared as a module but never
dispatched as a command, with #[allow(dead_code)] on all public
items. The valuable integration test (compute_changed_test_scope)
is relocated to core/extension/test/mod.rs where the function
lives. The is_test_path test was already covered by walker.rs.

refactor_tests.rs was a test file misplaced in the commands layer.
3 of 5 tests were exact duplicates of tests already in
core/refactor/plan/planner.rs. The 2 unique normalize_sources
tests are relocated there.

commands/ now contains only command primitives (thin CLI wrappers
around core APIs) and shared command utilities (utils/).

771 tests pass, cargo fmt clean, zero new warnings.
